+++ build.gradle
@@ -61,6 +61,7 @@ dependencies {
 	provided 'javax.servlet:javax.servlet-api:3.1.0'
 
 	testCompile "org.grails:grails-plugin-testing"
+	testCompile "org.grails.plugins:hibernate"
 	testCompile 'junit:junit:4.12'
 
 	console "org.grails:grails-console"
